+++ client/src/App.js
@@ -14,13 +14,10 @@ import ProfilePayment from "./pages/ProfilePayment";
 import ProfileDetails from "./pages/ProfileDetails";
 import PrivateRoute from "./pages/PrivateRoute";
 
-import Socket from "./pages/Socket";
-
 function App() {
   return (
     <MuiThemeProvider theme={theme}>
       <BrowserRouter>
-        <Route exact path="/socket" component={Socket} />
         <Route exact path="/" component={LandingPage} />
         <Route path="/login" component={LoginPage} />
         <Route path="/signup" component={SignUpPage} />
