+++ client/src/pages/Login.js
@@ -6,17 +6,35 @@ import Grid from "@material-ui/core/Grid";
 import "../App.scss";
 import Button from "@material-ui/core/Button";
 
+// import socket.io client
+import openSocket from 'socket.io-client';
+const socket = openSocket('http://localhost:3001');
+
+// test for socket.io
+function subscribeToTimer(cb) {
+  socket.on('timer', timestamp => cb(null, timestamp));
+  socket.emit('subscribeToTimer', 1000);
+}
+
 
 const initalState = {
   email: "",
   emailError: "",
   password: "",
   passwordError: "",
-  errors : ""  
+  errors : "",
+  timestamp: 'no timestamp yet'
 };
 
 class LoginPage extends Component {
   state = initalState;
+
+  constructor(props) {
+    super(props);
+    subscribeToTimer((err, timestamp) => this.setState({ 
+      timestamp 
+    }));
+  }
 
   handleEmailChange = event => {
     this.setState({ email: event.target.value });
@@ -85,6 +103,15 @@ class LoginPage extends Component {
                     <Grid item xs={12}>
                       <h1 className="center">LogIn</h1>
                     </Grid>
+                    
+                    {/* Testing for socket.io */}
+                    <div>
+                        <p>
+                            This is the timer value: {this.state.timestamp}
+                        </p>
+                    </div>
+
+
                     {
                       (this.state.errors)
                       ? <Grid item xs={12} className="pb-0 pt-0" style={{ color: "red" }}>
